src/settings.py
import logging
import pathlib

import environ

env = environ.Env()
BASE_DIR = pathlib.Path(__file__).parent.parent
environ.Env.read_env(str(BASE_DIR / ".env"))

DEBUG = env.bool("DEBUG", default=False)

logging.basicConfig(level=(logging.DEBUG if DEBUG else logging.INFO))

PG_CONN_STRING = env("PG_CONNECTION_STRING", default="")
