pyproject.toml
[project]
authors = [{ name = "Niccolò Cantù", email = "niccolo.cantu@nina.no" }]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
dependencies = [
  "django-environ",
  "click",
  "ibis-framework[duckdb,geospatial]",
  "rasterio",
  "pyarrow",
  "marimo",
]
description = ""
license = { text = "GPL-3.0+" }
name = "radar_data"
readme = "README.md"
requires-python = ">=3.10"
version = "0.1.0"

[project.optional-dependencies]
tools = ["pre-commit", "pyyaml", "GitPython"]

[project.scripts]
"radar_data_start" = "main:start"
"radar_data_extract_layer" = "extract_dataset:run_extract_dataset"
"radar_data_extract_points" = "extract_points:cli"

[tool.ruff]
fix = true

[tool.ruff.lint]
ignore = ["COM812"]
select = ["E", "W", "I", "F", "UP", "S", "B", "A", "COM", "LOG", "PTH", "Q"]

[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["E501", "ANN"]
"nina-python-init.py" = ["S603", "S607"]
